1. Added next scheduled time getter in schedules database provider in Mobile App 2. Implemented count down logic for scheduled time

+++ petfeed/lib/src/data/database/models/schedule.dart
@@ -30,7 +30,7 @@ class Schedule {
 
   Schedule.fromMap(Map<String, dynamic> map) {
     id = map[columnId];
-    amount = map[columnAmount];
+    amount = map[columnAmount].toDouble();
     day = map[columnDay];
     feedTime = DateTime.fromMillisecondsSinceEpoch(map[columnFeedTime].round());
     groupID = map[columnGroupId];
